# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AURUM Luxury Fashion eCommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  slug: string;
  name: string;
  brand: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  /** @nullable */
  badge?: string | null;
  colors: string[];
  sizes: string[];
  description?: string;
  figType: string;
  bgGradient: string;
  figColorA: string;
  figColorB: string;
  sku: string;
  stock: number;
  tags: string[];
}

export interface ProductSummary {
  total: number;
  clothing: number;
  shoes: number;
  accessories: number;
  sale: number;
  newArrivals: number;
}

export interface NewsletterInput {
  email: string;
}

export interface NewsletterResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProductsParams = {
category?: string;
badge?: string;
q?: string;
minPrice?: number;
maxPrice?: number;
sortBy?: ListProductsSortBy;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy];


export const ListProductsSortBy = {
  featured: 'featured',
  newest: 'newest',
  'price-asc': 'price-asc',
  'price-desc': 'price-desc',
  'name-asc': 'name-asc',
} as const;