# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AURUM Luxury Fashion eCommerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: newsletter
    description: Newsletter subscriptions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: badge
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [featured, newest, price-asc, price-desc, name-asc]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/summary:
    get:
      operationId: getProductSummary
      tags: [products]
      summary: Get product counts by category
      responses:
        "200":
          description: Category summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSummary"

  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterResponse"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required:
        - id
        - slug
        - name
        - brand
        - price
        - category
        - figType
        - bgGradient
        - figColorA
        - figColorB
        - sku
        - stock
        - sizes
        - colors
        - tags
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        category:
          type: string
        badge:
          type: ["string", "null"]
        colors:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        description:
          type: string
        figType:
          type: string
        bgGradient:
          type: string
        figColorA:
          type: string
        figColorB:
          type: string
        sku:
          type: string
        stock:
          type: integer
        tags:
          type: array
          items:
            type: string

    ProductSummary:
      type: object
      required:
        - total
        - clothing
        - shoes
        - accessories
        - sale
        - newArrivals
      properties:
        total:
          type: integer
        clothing:
          type: integer
        shoes:
          type: integer
        accessories:
          type: integer
        sale:
          type: integer
        newArrivals:
          type: integer

    NewsletterInput:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email

    NewsletterResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string