# lib/api-zod/src/generated/types/productSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AURUM Luxury Fashion eCommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductSummary {
  total: number;
  clothing: number;
  shoes: number;
  accessories: number;
  sale: number;
  newArrivals: number;
}