# lib/db/src/schema/products.ts
import { pgTable, text, integer, numeric, boolean, serial, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const productsTable = pgTable("products", {
  id: serial("id").primaryKey(),
  slug: text("slug").notNull().unique(),
  name: text("name").notNull(),
  brand: text("brand").notNull(),
  price: numeric("price", { precision: 10, scale: 2 }).notNull(),
  originalPrice: numeric("original_price", { precision: 10, scale: 2 }),
  category: text("category").notNull(),
  subcategory: text("subcategory"),
  badge: text("badge"),
  colors: text("colors").array().notNull().default([]),
  sizes: text("sizes").array().notNull().default([]),
  description: text("description"),
  figType: text("fig_type").notNull(),
  bgGradient: text("bg_gradient").notNull(),
  figColorA: text("fig_color_a").notNull(),
  figColorB: text("fig_color_b").notNull(),
  imageUrl: text("image_url"),
  images: text("images").array().notNull().default([]),
  featured: boolean("featured").notNull().default(false),
  sku: text("sku").notNull().unique(),
  stock: integer("stock").notNull().default(0),
  tags: text("tags").array().notNull().default([]),
  isActive: boolean("is_active").notNull().default(true),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }),
});

export const insertProductSchema = createInsertSchema(productsTable).omit({ id: true, createdAt: true });
export type InsertProduct = z.infer<typeof insertProductSchema>;
export type Product = typeof productsTable.$inferSelect;